Add a class for the QuestNav
This adds an easy to use robot side API for the quest which wraps the NT
subscriptions in a class. Open to feedback about what you'd like this
class to include.
It also switches the DriveSubsystem to use this class.

This will be the first of two PRs to add a published Java API for the
QuestNav.

Signed-off-by: Jade Turner <spacey-sooty@proton.me>